Move PlayerRepository to allow making Player constructor protected
