+++ bukkit/src/main/java/de/t14d3/zones/commands/CancelCommand.java
@@ -3,8 +3,8 @@ package de.t14d3.zones.commands;
 import de.t14d3.zones.BukkitPlatform;
 import de.t14d3.zones.RegionManager;
 import de.t14d3.zones.ZonesBukkit;
+import de.t14d3.zones.objects.PlayerRepository;
 import de.t14d3.zones.utils.Messages;
-import de.t14d3.zones.utils.PlayerRepository;
 import dev.jorel.commandapi.CommandAPICommand;
 import net.kyori.adventure.text.minimessage.MiniMessage;
 import org.bukkit.World;
