+++ bukkit/src/main/java/de/t14d3/zones/commands/SelectCommand.java
@@ -5,9 +5,9 @@ import de.t14d3.zones.RegionKey;
 import de.t14d3.zones.RegionManager;
 import de.t14d3.zones.ZonesBukkit;
 import de.t14d3.zones.objects.BlockLocation;
+import de.t14d3.zones.objects.PlayerRepository;
 import de.t14d3.zones.objects.World;
 import de.t14d3.zones.utils.Messages;
-import de.t14d3.zones.utils.PlayerRepository;
 import dev.jorel.commandapi.BukkitTooltip;
 import dev.jorel.commandapi.CommandAPICommand;
 import dev.jorel.commandapi.StringTooltip;
