+++ bukkit/src/main/java/de/t14d3/zones/listeners/PlayerQuitListener.java
@@ -1,8 +1,8 @@
 package de.t14d3.zones.listeners;
 
 import de.t14d3.zones.Zones;
+import de.t14d3.zones.objects.PlayerRepository;
 import de.t14d3.zones.permissions.CacheUtils;
-import de.t14d3.zones.utils.PlayerRepository;
 import org.bukkit.event.EventHandler;
 import org.bukkit.event.Listener;
 import org.bukkit.event.player.PlayerQuitEvent;
