+++ fabric/src/main/java/de/t14d3/zones/fabric/commands/SetCommand.java
@@ -9,9 +9,9 @@ import de.t14d3.zones.RegionManager;
 import de.t14d3.zones.ZonesFabric;
 import de.t14d3.zones.objects.Flag;
 import de.t14d3.zones.objects.Player;
+import de.t14d3.zones.objects.PlayerRepository;
 import de.t14d3.zones.permissions.flags.Flags;
 import de.t14d3.zones.utils.Messages;
-import de.t14d3.zones.utils.PlayerRepository;
 import me.lucko.fabric.api.permissions.v0.Permissions;
 import net.kyori.adventure.platform.modcommon.impl.NonWrappingComponentSerializer;
 import net.kyori.adventure.platform.modcommon.impl.WrappedComponent;
